+++ src/server/routes/index.ts
@@ -2,8 +2,8 @@ import express from 'express';
 import { adapters, H5PEditor } from 'h5p-nodejs-library';
 
 import h5pConfig from '../../config/h5pConfig';
-import lumiH5PRoutes from './lumi-h5p';
-import trackRoutes from './track';
+import lumiRoutes from './lumiRoutes';
+import trackingRoutes from './trackingRoutes';
 
 import Logger from '../helpers/Logger';
 import User from '../h5pImplementations/User';
@@ -14,7 +14,7 @@ export default function (h5pEditor: H5PEditor): express.Router {
 
     log.info('setting up routes');
 
-    router.use('/api/track/v0', trackRoutes());
+    router.use('/api/track/v0', trackingRoutes());
 
     // Adding dummy user to make sure all requests can be handled
     router.use((req, res, next) => {
@@ -31,7 +31,7 @@ export default function (h5pEditor: H5PEditor): express.Router {
         )
     );
 
-    router.use('/api/lumi-h5p/v1', lumiH5PRoutes(h5pEditor));
+    router.use('/api/lumi-h5p/v1', lumiRoutes(h5pEditor));
 
     router.get('*', express.static(`${__dirname}/../../client`));
